// src/data/utils/interfaces.ts
export interface SearchSuggestion {
  title: string;
  subtitle: string;
  lat?: number;
  long?: number;
  stopCode?: string;
  placeId?: string;
  type: 'stop' | 'map' | 'my-location';
}